Created persistence and models modules
