+++ botpersistence/src/main/java/com/blackbook/botpersistence/dao/PromotionsRepository.java
@@ -1,6 +1,7 @@
-package com.blackbook.botrest.dao;
+package com.blackbook.botpersistence.dao;
 
-import com.blackbook.dao.model.Promotion;
+
+import com.blackbook.botpersistence.model.Promotion;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
