+++ botpersistence/src/main/resources/db.migration/V1_01__Prices_Bookstores_Books_Authors.sql
