+++ botrest/src/main/java/com/blackbook/botrest/ApiCodes.java
@@ -1,0 +1,7 @@
+package com.blackbook.botrest;
+
+public interface ApiCodes {
+
+    String PERSISTENCE_API = "http://localhost:8101";
+
+}
