+++ botrest/src/main/java/com/blackbook/botrest/BotApplication.java
@@ -2,11 +2,13 @@ package com.blackbook.botrest;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
+@EntityScan(basePackages = "com.blackbook.dao")
 @SpringBootApplication
 public class BotApplication {
