+++ botrest/src/main/java/com/blackbook/botrest/controller/PromotionsController.java
@@ -1,29 +1,35 @@
 package com.blackbook.botrest.controller;
 
-import com.blackbook.botrest.model.view.promotion.PromotionView;
-import com.blackbook.botrest.dao.PromotionsRepository;
+import com.blackbook.botpersistence.model.Promotion;
+import com.blackbook.botrest.ApiCodes;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.type.TypeFactory;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestTemplate;
+import view.promotion.PromotionView;
 
+import java.io.IOException;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @RestController
 @RequestMapping(value = "/api/promotions")
 public class PromotionsController {
 
-    @Autowired
-    private PromotionsRepository promotionsRepository;
+    @RequestMapping(method = RequestMethod.GET)
+    public List<PromotionView> getPromotions() throws IOException {
+        RestTemplate restTemplate = new RestTemplate();
 
-    public List<PromotionView> getPromotions() {
+        ResponseEntity<String> response
+                = restTemplate.getForEntity(ApiCodes.PERSISTENCE_API + "/api/promotions", String.class);
+        ObjectMapper mapper = new ObjectMapper();
+        List<PromotionView> list = mapper.readValue(response.getBody(), TypeFactory.defaultInstance().constructCollectionType(List.class, PromotionView.class));
 
-        return promotionsRepository
-                .findAll()
-                .stream()
-                .map(PromotionView::fromPromotion)
-                .collect(Collectors.toList());
+        return list;
     }
+
 }
