+++ botrest/src/main/resources/application.yml
@@ -1,0 +1,2 @@
+server:
+  port: 8080
