+++ models/src/main/java/ModelsApplication.java
@@ -1,0 +1,10 @@
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModelsApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ModelsApplication.class);
+    }
+
+}
