+++ models/src/main/java/view/promotion/BookAuthorView.java
@@ -1,0 +1,12 @@
+package view.promotion;
+
+import lombok.Data;
+
+/**
+ * Created by tymek on 25.08.17.
+ */
+@Data
+public class BookAuthorView {
+    private String name;
+    private String surname;
+}
