+++ models/src/main/java/view/promotion/BookstoreView.java
@@ -1,0 +1,10 @@
+package view.promotion;
+
+import lombok.Data;
+
+@Data
+public class BookstoreView {
+    private long id;
+    private String name;
+    private String details;
+}
